requirements.txt
absl-py==1.3.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
aiohttp==3.8.3 ; python_version >= "3.8" and python_full_version < "3.9.7"
aiosignal==1.2.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
altair==4.2.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
async-timeout==4.0.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
attrs==22.1.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
blinker==1.5 ; python_version >= "3.8" and python_full_version < "3.9.7"
cachetools==5.2.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
certifi==2022.9.24 ; python_version >= "3.8" and python_full_version < "3.9.7"
charset-normalizer==2.1.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
click==8.1.3 ; python_version >= "3.8" and python_full_version < "3.9.7"
colorama==0.4.6 ; python_version >= "3.8" and python_full_version < "3.9.7" and platform_system == "Windows"
commonmark==0.9.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
decorator==5.1.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
entrypoints==0.4 ; python_version >= "3.8" and python_full_version < "3.9.7"
frozenlist==1.3.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
fsspec[http]==2022.10.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
gitdb==4.0.9 ; python_version >= "3.8" and python_full_version < "3.9.7"
gitpython==3.1.29 ; python_version >= "3.8" and python_full_version < "3.9.7"
google-auth-oauthlib==0.4.6 ; python_version >= "3.8" and python_full_version < "3.9.7"
google-auth==2.13.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
grpcio==1.50.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
idna==3.4 ; python_version >= "3.8" and python_full_version < "3.9.7"
importlib-metadata==5.0.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
importlib-resources==5.10.0 ; python_version >= "3.8" and python_version < "3.9"
jinja2==3.1.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
joblib==1.2.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
jsonschema==4.16.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
markdown==3.4.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
markupsafe==2.1.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
more-itertools==9.0.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
multidict==6.0.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
numpy==1.23.4 ; python_version >= "3.8" and python_full_version < "3.9.7"
oauthlib==3.2.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
packaging==21.3 ; python_version >= "3.8" and python_full_version < "3.9.7"
pandas==1.5.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
pillow==9.3.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
protobuf==3.19.6 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyarrow==10.0.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyasn1-modules==0.2.8 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyasn1==0.4.8 ; python_version >= "3.8" and python_full_version < "3.9.7"
pydeck==0.8.0b4 ; python_version >= "3.8" and python_full_version < "3.9.7"
pydeprecate==0.3.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
pygments==2.13.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
pympler==1.0.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyparsing==3.0.9 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyrsistent==0.18.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
pytorch-lightning==1.7.7 ; python_version >= "3.8" and python_full_version < "3.9.7"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.8" and python_full_version < "3.9.7"
pytz==2022.5 ; python_version >= "3.8" and python_full_version < "3.9.7"
pyyaml==6.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
rdkit-pypi==2022.9.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
requests-oauthlib==1.3.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
requests==2.28.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
rich==12.6.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
rsa==4.9 ; python_version >= "3.8" and python_full_version < "3.9.7"
scikit-learn==1.1.3 ; python_version >= "3.8" and python_full_version < "3.9.7"
scipy==1.9.3 ; python_version >= "3.8" and python_full_version < "3.9.7"
semver==2.13.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
setuptools==65.5.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
six==1.16.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
smmap==5.0.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
streamlit==1.14.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
tensorboard-data-server==0.6.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
tensorboard-plugin-wit==1.8.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
tensorboard==2.10.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
threadpoolctl==3.1.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
toml==0.10.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
toolz==0.12.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
torch-geometric==2.1.0.post1 ; python_version >= "3.8" and python_full_version < "3.9.7"
torch==1.12.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
torchmetrics==0.10.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
tornado==6.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
tqdm==4.64.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
tzdata==2022.6 ; python_version >= "3.8" and python_full_version < "3.9.7"
tzlocal==4.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
urllib3==1.26.12 ; python_version >= "3.8" and python_full_version < "3.9.7"
validators==0.20.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
watchdog==2.1.9 ; python_version >= "3.8" and python_full_version < "3.9.7" and platform_system != "Darwin"
werkzeug==2.2.2 ; python_version >= "3.8" and python_full_version < "3.9.7"
wheel==0.37.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
yarl==1.8.1 ; python_version >= "3.8" and python_full_version < "3.9.7"
zipp==3.10.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
cloudpickle==2.2.0 ; python_version >= "3.8" and python_full_version < "3.9.7"
--find-links https://data.pyg.org/whl/torch-1.12.0+cpu.html
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric